refactor: PathService로부터 PathBuilder 클래스를 분리
